Rename request -> requests, standard code fixes
